Move tests for POST /ctms (create a record)
