+++ tests/unit/conftest.py
@@ -1,4 +1,5 @@
 """pytest fixtures for the CTMS app"""
+from typing import Callable, Optional
 from uuid import UUID
 
 import pytest
@@ -10,9 +11,18 @@ from sqlalchemy_utils.functions import create_database, database_exists, drop_da
 
 from ctms.app import app, get_api_client, get_db
 from ctms.config import Settings
-from ctms.crud import create_api_client, create_contact
+from ctms.crud import (
+    create_api_client,
+    create_contact,
+    get_amo_by_email_id,
+    get_contacts_by_any_id,
+    get_fxa_by_email_id,
+    get_mofo_by_email_id,
+    get_newsletters_by_email_id,
+    get_vpn_by_email_id,
+)
 from ctms.models import Base
-from ctms.schemas import ApiClientSchema
+from ctms.schemas import ApiClientSchema, ContactSchema
 from tests.unit.sample_data import SAMPLE_CONTACTS
 
 
@@ -159,3 +169,87 @@ def client_id_and_secret(dbsession):
     create_api_client(dbsession, api_client, secret)
     dbsession.flush()
     return (api_client.client_id, secret)
+
+
+@pytest.fixture
+def post_contact(request, client, dbsession):
+    _id = (
+        request.param
+        if hasattr(request, "param")
+        else "d1da1c99-fe09-44db-9c68-78a75752574d"
+    )
+    email_id = UUID(str(_id))
+    contact = SAMPLE_CONTACTS[email_id]
+    fields_not_written = SAMPLE_CONTACTS.get_not_written(email_id)
+
+    def _add(
+        modifier: Callable[[ContactSchema], ContactSchema] = lambda x: x,
+        code: int = 201,
+        stored_contacts: int = 1,
+        check_redirect: bool = True,
+        query_fields: Optional[dict] = None,
+        check_written: bool = True,
+    ):
+        if query_fields is None:
+            query_fields = {"primary_email": contact.email.primary_email}
+        sample = contact.copy(deep=True)
+        sample = modifier(sample)
+        resp = client.post("/ctms", sample.json())
+        assert resp.status_code == code, resp.text
+        if check_redirect:
+            assert resp.headers["location"] == f"/ctms/{sample.email.email_id}"
+        saved = [
+            ContactSchema(**c)
+            for c in get_contacts_by_any_id(dbsession, **query_fields)
+        ]
+        assert len(saved) == stored_contacts
+
+        # Now make sure that we skip writing default models
+        def _check_written(field, getter, result_list=False):
+            # We delete this field in one test case so we have to check
+            # to see if it is even there
+            if hasattr(sample.email, "email_id") and sample.email.email_id is not None:
+                written_id = sample.email.email_id
+            else:
+                written_id = resp.headers["location"].split("/")[-1]
+            results = getter(dbsession, written_id)
+            if sample.dict().get(field) and code in {200, 201}:
+                if field in fields_not_written:
+                    if result_list:
+                        assert (
+                            results == []
+                        ), f"{email_id} has field `{field}` but it is _default_ and it should _not_ have been written to db"
+                    else:
+                        assert (
+                            results is None
+                        ), f"{email_id} has field `{field}` but it is _default_ and it should _not_ have been written to db"
+                else:
+                    assert (
+                        results
+                    ), f"{email_id} has field `{field}` and it should have been written to db"
+            else:
+                if result_list:
+                    assert (
+                        results == []
+                    ), f"{email_id} does not have field `{field}` and it should _not_ have been written to db"
+                else:
+                    assert (
+                        results is None
+                    ), f"{email_id} does not have field `{field}` and it should _not_ have been written to db"
+
+        if check_written:
+            _check_written("amo", get_amo_by_email_id)
+            _check_written("fxa", get_fxa_by_email_id)
+            _check_written("mofo", get_mofo_by_email_id)
+            _check_written("newsletters", get_newsletters_by_email_id, result_list=True)
+            _check_written("vpn_waitlist", get_vpn_by_email_id)
+
+        # Check that GET returns the same contact
+        if code in {200, 201}:
+            dbsession.expunge_all()
+            get_resp = client.get(resp.headers["location"])
+            assert resp.json() == get_resp.json()
+
+        return saved, sample, email_id
+
+    return _add
