+++ tests/unit/test_api.py
@@ -81,90 +81,6 @@ def test_authorized_api_call_succeeds(client, example_contact, method, path, par
             assert resp.status_code == 200
 
 
-@pytest.fixture
-def post_contact(request, client, dbsession):
-    _id = (
-        request.param
-        if hasattr(request, "param")
-        else "d1da1c99-fe09-44db-9c68-78a75752574d"
-    )
-    email_id = UUID(str(_id))
-    contact = SAMPLE_CONTACTS[email_id]
-    fields_not_written = SAMPLE_CONTACTS.get_not_written(email_id)
-
-    def _add(
-        modifier: Callable[[ContactSchema], ContactSchema] = lambda x: x,
-        code: int = 201,
-        stored_contacts: int = 1,
-        check_redirect: bool = True,
-        query_fields: Optional[dict] = None,
-        check_written: bool = True,
-    ):
-        if query_fields is None:
-            query_fields = {"primary_email": contact.email.primary_email}
-        sample = contact.copy(deep=True)
-        sample = modifier(sample)
-        resp = client.post("/ctms", sample.json())
-        assert resp.status_code == code, resp.text
-        if check_redirect:
-            assert resp.headers["location"] == f"/ctms/{sample.email.email_id}"
-        saved = [
-            ContactSchema(**c)
-            for c in get_contacts_by_any_id(dbsession, **query_fields)
-        ]
-        assert len(saved) == stored_contacts
-
-        # Now make sure that we skip writing default models
-        def _check_written(field, getter, result_list=False):
-            # We delete this field in one test case so we have to check
-            # to see if it is even there
-            if hasattr(sample.email, "email_id") and sample.email.email_id is not None:
-                written_id = sample.email.email_id
-            else:
-                written_id = resp.headers["location"].split("/")[-1]
-            results = getter(dbsession, written_id)
-            if sample.dict().get(field) and code in {200, 201}:
-                if field in fields_not_written:
-                    if result_list:
-                        assert (
-                            results == []
-                        ), f"{email_id} has field `{field}` but it is _default_ and it should _not_ have been written to db"
-                    else:
-                        assert (
-                            results is None
-                        ), f"{email_id} has field `{field}` but it is _default_ and it should _not_ have been written to db"
-                else:
-                    assert (
-                        results
-                    ), f"{email_id} has field `{field}` and it should have been written to db"
-            else:
-                if result_list:
-                    assert (
-                        results == []
-                    ), f"{email_id} does not have field `{field}` and it should _not_ have been written to db"
-                else:
-                    assert (
-                        results is None
-                    ), f"{email_id} does not have field `{field}` and it should _not_ have been written to db"
-
-        if check_written:
-            _check_written("amo", get_amo_by_email_id)
-            _check_written("fxa", get_fxa_by_email_id)
-            _check_written("mofo", get_mofo_by_email_id)
-            _check_written("newsletters", get_newsletters_by_email_id, result_list=True)
-            _check_written("vpn_waitlist", get_vpn_by_email_id)
-
-        # Check that GET returns the same contact
-        if code in {200, 201}:
-            dbsession.expunge_all()
-            get_resp = client.get(resp.headers["location"])
-            assert resp.json() == get_resp.json()
-
-        return saved, sample, email_id
-
-    return _add
-
-
 def _compare_written_contacts(
     contact,
     sample,
@@ -186,112 +102,6 @@ def _compare_written_contacts(
         setattr(sample, f, [] if f == "newsletters" else None)
 
     assert saved_contact.idempotent_equal(sample)
-
-
-@pytest.mark.parametrize("post_contact", SAMPLE_CONTACTS.keys(), indirect=True)
-def test_create_basic_no_id(post_contact):
-    """Most straightforward contact creation succeeds when email_id is not a key."""
-
-    def _remove_id(contact):
-        del contact.email.email_id
-        return contact
-
-    saved_contacts, sample, email_id = post_contact(
-        modifier=_remove_id, check_redirect=False
-    )
-    _compare_written_contacts(
-        saved_contacts[0], sample, email_id, ids_should_be_identical=False
-    )
-
-
-@pytest.mark.parametrize("post_contact", SAMPLE_CONTACTS.keys(), indirect=True)
-def test_create_basic_id_is_none(post_contact):
-    """Most straightforward contact creation succeeds when email_id is None."""
-
-    def _remove_id(contact):
-        contact.email.email_id = None
-        return contact
-
-    saved_contacts, sample, email_id = post_contact(
-        modifier=_remove_id, check_redirect=False
-    )
-    _compare_written_contacts(
-        saved_contacts[0], sample, email_id, ids_should_be_identical=False
-    )
-
-
-@pytest.mark.parametrize("post_contact", SAMPLE_CONTACTS.keys(), indirect=True)
-def test_create_basic_with_id(post_contact):
-    """Most straightforward contact creation succeeds when email_id is specified."""
-    saved_contacts, sample, email_id = post_contact()
-    _compare_written_contacts(saved_contacts[0], sample, email_id)
-
-
-@pytest.mark.parametrize("post_contact", SAMPLE_CONTACTS.keys(), indirect=True)
-def test_create_basic_idempotent(post_contact):
-    """Creating a contact works across retries."""
-    saved_contacts, sample, email_id = post_contact()
-    _compare_written_contacts(saved_contacts[0], sample, email_id)
-    saved_contacts, _, _ = post_contact(code=200)
-    _compare_written_contacts(saved_contacts[0], sample, email_id)
-
-
-@pytest.mark.parametrize("post_contact", SAMPLE_CONTACTS.keys(), indirect=True)
-def test_create_basic_with_id_collision(post_contact):
-    """Creating a contact with the same id but different data fails."""
-    _, sample, _ = post_contact()
-
-    def _change_mailing(contact):
-        assert contact.email.mailing_country != "mx", "sample data has changed"
-        contact.email.mailing_country = "mx"
-        return contact
-
-    # We set check_written to False because the rows it would check for normally
-    # are actually here due to the first write
-    saved_contacts, _, _ = post_contact(
-        modifier=_change_mailing, code=409, check_redirect=False, check_written=False
-    )
-    assert saved_contacts[0].email.mailing_country == sample.email.mailing_country
-
-
-@pytest.mark.parametrize("post_contact", SAMPLE_CONTACTS.keys(), indirect=True)
-def test_create_basic_with_basket_collision(post_contact):
-    """Creating a contact with diff ids but same email fails.
-    We override the basket token so that we know we're not colliding on that here.
-    See test_create_basic_with_email_collision below for that check
-    """
-    saved_contacts, orig_sample, email_id = post_contact()
-    _compare_written_contacts(saved_contacts[0], orig_sample, email_id)
-
-    def _change_basket(contact):
-        contact.email.email_id = UUID("229cfa16-a8c9-4028-a9bd-fe746dc6bf73")
-        contact.email.basket_token = UUID("df9f7086-4949-4b2d-8fcf-49167f8f783d")
-        return contact
-
-    saved_contacts, _, _ = post_contact(
-        modifier=_change_basket, code=409, check_redirect=False
-    )
-    _compare_written_contacts(saved_contacts[0], orig_sample, email_id)
-
-
-@pytest.mark.parametrize("post_contact", SAMPLE_CONTACTS.keys(), indirect=True)
-def test_create_basic_with_email_collision(post_contact):
-    """Creating a contact with diff ids but same basket token fails.
-    We override the email so that we know we're not colliding on that here.
-    See other test for that check
-    """
-    saved_contacts, orig_sample, email_id = post_contact()
-    _compare_written_contacts(saved_contacts[0], orig_sample, email_id)
-
-    def _change_primary_email(contact):
-        contact.email.email_id = UUID("229cfa16-a8c9-4028-a9bd-fe746dc6bf73")
-        contact.email.primary_email = "foo@example.com"
-        return contact
-
-    saved_contacts, _, _ = post_contact(
-        modifier=_change_primary_email, code=409, check_redirect=False
-    )
-    _compare_written_contacts(saved_contacts[0], orig_sample, email_id)
 
 
 @pytest.fixture
